refactor(frontend): split speakers page into tab components
Phase 5.2 — no-behavior/no-visual split. The route page stays the coordinator (data
loading, all cluster/profile/inbox action handlers, merge/split/unassign state, WebSocket
clustering listeners, global keyboard handler, the 3 modals) and tab switcher; the tab
bodies + floating player move to thin presentational children (2269 -> 1169):
- speakers/ClustersTab.svelte (591) — toolbar + progress + identified/unidentified sections
  (renders the unchanged SpeakerClusterCard/ClusterMemberList) + pagination.
- speakers/ProfilesTab.svelte (399) — profile grid, gender-confirm, avatar upload/edit.
- speakers/InboxTab.svelte (128) — inbox list + pagination.
- speakers/SpeakerPreviewPlayer.svelte (208) — floating Plyr mini-player (bind:playerRef).
Each receives data + mode state as props and dispatches user actions up; no API/store logic
moved down. Kept the existing tab-button markup (the E2E locates tabs via has-text), did NOT
adopt ui/Tabs. SpeakerClusterCard/ClusterMemberList untouched.

Verified: svelte-check 0/0 (856 files), build, 52 unit tests; 21/23 speaker E2E pass. The 2
failing assertions (.gender-icon expects the symbol but the dev data shows a confirmed ✓ on
the first cluster) reproduce IDENTICALLY on the pre-split code (verified via stash) — they are
pre-existing/data-dependent, not caused by this refactor.
